:white_check_mark: Add xy visualization test for thresholds
